pyproject.toml
[project]
name = "smart-job-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "modal>=1.4.2",
    "ollama>=0.6.1",
    "pdfplumber>=0.11.9",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "uvicorn>=0.46.0",
    "websockets>=16.0",
]
